requirements.txt
ase==3.21.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
configparser==5.0.1
cycler==0.10.0
decorator==4.4.2
docker-pycreds==0.4.0
gitdb==4.0.5
GitPython==3.1.13
googledrivedownloader==0.4
h5py==3.1.0
idna==2.10
isodate==0.6.0
Jinja2==2.11.3
joblib==1.0.0
kiwisolver
llvmlite==0.35.0
MarkupSafe==1.1.1
matplotlib
networkx==2.5
numba==0.52.0
numpy~=1.19.5
olefile
pandas==1.2.1
pathtools==0.1.2
Pillow
promise==2.3
protobuf==3.14.0
psutil==5.8.0
PubChemPy==1.0.4
pycairo==1.20.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-louvain==0.15
pytz
PyYAML==5.4.1
rdflib==5.0.0
reportlab==3.5.60
requests==2.25.1
scikit-learn==0.24.1
scipy==1.6.0
sentry-sdk==0.20.2
shortuuid==1.0.1
six
smmap==3.0.5
SQLAlchem
subprocess32==3.5.4
threadpoolctl==2.1.0
torch==1.7.1
torch-geometric
torch-scatter==2.0.5
torch-sparse==0.6.8
torch-spline-conv==1.2.0
tornado
tqdm==4.56.0
typing-extensions==3.7.4.3
urllib3==1.26.3
wandb~=0.10.21
rdkit~=2020.09.4